Refactor simulation functions to classes
Rewrite the functions to wrap the surveillance functions as classes
to follow a similar API as scikit-learn where a model is instantiated
with some parameters. The simulations are then elicited executing
the .simulate() method on the model object which itself takes parameters.
To make things clean, I added an base class to enforce the idea of an API.
The tests and the tests data was adapted to these changes.

I fixed an occurence where a column of the simulations was still
named raw_cases instead of n_cases. I wrote a test to check this
so that this mistakes does not repeat.

Furthermore, I updated the docstrings which still concluded paramter names
of the R functions instead of the new Python wrapper. Finally, I adapted
the default parameters to their type hint (1 -> 1.), corrected unprecise
type hints, and avoided using the same variable names for several steps
of data processing.