Add wrapper for simulation algos from Surveillance  (#21)
* Add wrapper for surveillance's endemic, epidemic, and hhh time series simulation

* Add column to seasonal_noise and Black format code

* Delete: hhh simulation since not usable without algo.hhh

* Refactor: using type hints, extract repeated functions, make frame trainable

Create a utils with functions to convert R lists to pandas.DataFrame
and add dates to make simulated data usable by the outbreak detection algorithms.

To make point_source simulation compatible with outbreak detection algorithms
I converted all cases to outbreak cases if there where some. After some testing
I found that mostly 8/9 of the cases in a outbreak are outbreak cases with the
default parameters.

Add some tests for the utils and rename function parameters.

* Refactor directory name

* Add tests to simulation functions and more type hints

* Fix wrong module name (simulation -> simulations)

* Refactor module name (simulations -> simulation)

* Refactor simulation functions to classes

Rewrite the functions to wrap the surveillance functions as classes
to follow a similar API as scikit-learn where a model is instantiated
with some parameters. The simulations are then elicited executing
the .simulate() method on the model object which itself takes parameters.
To make things clean, I added an base class to enforce the idea of an API.
The tests and the tests data was adapted to these changes.

I fixed an occurence where a column of the simulations was still
named raw_cases instead of n_cases. I wrote a test to check this
so that this mistakes does not repeat.

Furthermore, I updated the docstrings which still concluded paramter names
of the R functions instead of the new Python wrapper. Finally, I adapted
the default parameters to their type hint (1 -> 1.), corrected unprecise
type hints, and avoided using the same variable names for several steps
of data processing.

* Fix linter error caused by typehints

Using a typehint for an object where its reference is not known in the module alone
like:

	def return_frame(df) -> pandas.DataFrame:
		return df
I included the respective imports in the beginning of the module.
Furthermore, I fixed a type error in the base .simulate() method.

* Fix type annotations and imports

Since the return type was set in quotation marks (e.g -> "pandas.DataFrame")
LGTM made the alert that there are unused imports. Thus I changed the
type anotation by removing the quotation makrs. Furthermore, there
was indeed a simply unused import statement left that this commit removes.

* Fix incompatible type signatures

* Use abs in instead of abcmeta

* Rename and change to dataclasses

* Add missing base class

* Fix: wrong default parameter

* Fix wrong default parameter