Add a simulation based on the negative binomial distribution (Python only) (#27)
* Add a simulation based on the negative binomial distribution and fix simulation documentation

The model is taken form the paper "An improved algorithm for outbreak detection in multiple surveillance system"

* Refactor based on pre-commit hook

* Fix docstring formatting errors, harmonize param names, and add test

* Refactor output of DataFrames and replace for loops with numpy

* Refactor dosctring format and add more descriptive parameter name

* Remove n_outbreak_cases column if not needed

* Remove outbreak labels from Point Source simulation

Outbreak labels are removed because we cannot say whether
the observed case counts are based on an current outbreak
or are just endemic numbers. The currently used algorithm
would need to be adjusted to allow the retrieval of the
oubtreak information.