Adopt code to pandas 1.x
With the newest change in pandas, the frequency
of a DateTimeIndex is not automatically inferred
as expected by the previous code.
This PR makes the inferrence of the DateTimeIndex
frequency explicit.

Furthermore, it moves one function to read in files
to tests/utils.py where together with a similar
function.