Update libraries due to breaking versions.
Major version in sphinx-bibtex is incompatible with this library's
sphinx version. This commit restricts the version number:
https://github.com/executablebooks/jupyter-book/issues/1137#issuecomment-743994864

The latest release of jinja2 introduces was not compatible with
this library's sphinx version. This commit restricts the version number:
https://github.com/readthedocs/readthedocs.org/issues/9038#issuecomment-1078927910

isort is used in pre-commit hook but does not come with the
requirements-dev.txt. This commit adds it to have it available during
development.

Building the docs with a Jupyter notebook in it requires pandoc. The
README is updated to include an installation instruction. Using
conda is suggested in this stackoverflow question:
https://stackoverflow.com/questions/62398231/building-docs-fails-due-to-missing-pandoc