Fix failing tests and formatting in simulations.
test_simulations.py was throwing dtype errors since sometimes, n_cases
was int_32 and sometimes int_64. With this commit, data type is casted
to pass the unit test.

Imports were not sorted correctly according to isort which was fixed.

Docstring of simulate was miswritten and missed a "s". This commit
fixes it.